requirements.txt
langchain
langchain-community
langchain-core
langchain-text-splitters
zhipuai
chromadb
pydantic
python-dotenv
sentence-transformers
rank_bm25
fastapi
uvicorn
python-multipart
httpx
loguru
